+++ src/views/rules/RulesMagicAbilities.vue
@@ -136,5 +136,3 @@ export default Vue.extend({
   },
 });
 </script>
-
-<style></style>